started fixing some bias.
made the sign stuff random when a root is not "for sure" found. fixed a problem with the angles not being centered. fixed a thing with the tangent lines not being fixed. fixed a problem where changing the ray direction was changing the bias (when it shouldn't). 
there is still some bias and it is unclear where it is coming from.